+++ src/app/modules/auth/auth.routes.ts
@@ -2,7 +2,6 @@ import express, { NextFunction, Request, Response } from 'express'
 import { AuthController } from './auth.controller'
 import { AuthValidation } from './authValidation'
 import validateRequest from '../../middlewares/validateRequest'
-import { upload } from '../../middlewares/multerMiddleware'
 
 const router = express.Router()
 
@@ -18,15 +17,5 @@ router.post(
   AuthController.loginUser,
 )
 router.get('/refresh-token', AuthController.refreshToken)
-router.delete('/:id', AuthController.deleteUser)
-
-router.patch(
-  '/profile/:id',
-  upload.fields([{ name: 'avatar', maxCount: 1 }]),
-  (req: Request, res: Response, next: NextFunction) => {
-    req.body = AuthValidation.updateUserZodSchema.parse(req.body)
-    return AuthController.updateUser(req, res, next)
-  },
-)
 
 export const AuthRoutes = router
