+++ src/app/modules/user/user.interface.ts
@@ -1,0 +1,16 @@
+export type IWhereClause = {
+  OR?: Array<{
+    firstName?: {
+      contains: string | undefined
+      mode: 'insensitive'
+    }
+    address?: {
+      contains: string | undefined
+      mode: 'insensitive'
+    }
+    email?: {
+      contains?: string | undefined
+      mode: 'insensitive'
+    }
+  }>
+}
