+++ src/app/routes/index.ts
@@ -1,5 +1,6 @@
 import express from 'express'
 import { AuthRoutes } from '../modules/auth/auth.routes'
+import { UserRoutes } from '../modules/user/user.routes'
 
 const router = express.Router()
 
@@ -7,6 +8,10 @@ const moduleRoutes = [
   {
     path: '/auth',
     route: AuthRoutes,
+  },
+  {
+    path: '/user',
+    route: UserRoutes,
   },
 ]
